# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evaluaciones OMR API",
    "description": "Servicio de evaluaciones con reconocimiento óptico de marcas (OMR).\n\n**Flujo principal**\n1. Crear una evaluación (banco de preguntas) → agregar secciones → cargar preguntas con `seccion_id`.\n2. Generar hojas: `POST /generar/{evaluacion_id}` con una selección **template** (reparto por\n   cantidad/porcentaje por sección) o **estructurada** (banco completo o subconjunto ordenado).\n   El `identificador` es opcional: si falta se autogenera (`{short_id}-{n}`) y la reimpresión\n   con el mismo identificador devuelve la misma variante.\n3. Escanear la hoja llena y evaluarla con `POST /evaluar` (califica contra el banco).\n\n**Hoja genérica (sin examen)**\n- `POST /generar/generica` genera solo burbujas (N preguntas × M opciones, 2–5).\n- `POST /leer` devuelve lo llenado (identificador persona + burbujas marcadas) sin contrastar nada.\n\nVer **Redoc** (`/redoc`) para la referencia completa o el esquema OpenAPI en `/openapi.json`.",
    "contact": {
      "name": "Evaluaciones OMR"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/generar/generica": {
      "post": {
        "tags": [
          "Hoja genérica y lectura"
        ],
        "summary": "Generar Hoja Generica Endpoint",
        "operationId": "generar_hoja_generica_endpoint_api_v1_generar_generica_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerarGenericaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerarGenericaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/generica/{identificador}": {
      "get": {
        "tags": [
          "Hoja genérica y lectura"
        ],
        "summary": "Obtener Hoja Generica",
        "description": "Recupera una hoja genérica por identificador (PDF en base64, reimpresión).",
        "operationId": "obtener_hoja_generica_api_v1_generica__identificador__get",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerarGenericaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Hoja genérica y lectura"
        ],
        "summary": "Eliminar Hoja Generica",
        "description": "Elimina una hoja genérica y sus archivos de storage.",
        "operationId": "eliminar_hoja_generica_api_v1_generica__identificador__delete",
        "parameters": [
          {
            "name": "identificador",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leer": {
      "post": {
        "tags": [
          "Hoja genérica y lectura"
        ],
        "summary": "Leer Hoja Generica",
        "description": "Lee una hoja genérica escaneada y devuelve lo llenado (sin contrastar).",
        "operationId": "leer_hoja_generica_api_v1_leer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LecturaGenericaLoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvaluarRequest": {
        "properties": {
          "pdf_escaneado_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Escaneado Base64"
          },
          "imagen_escaneada_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagen Escaneada Base64"
          },
          "identificador": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identificador"
          }
        },
        "type": "object",
        "title": "EvaluarRequest"
      },
      "GenerarGenericaRequest": {
        "properties": {
          "identificador": {
            "type": "string",
            "title": "Identificador",
            "description": "Identificador de la hoja genérica (único; la reimpresión con el mismo id devuelve la misma hoja)"
          },
          "cantidad_preguntas": {
            "type": "integer",
            "title": "Cantidad Preguntas",
            "description": "Número de preguntas (filas de burbujas)"
          },
          "opciones": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Opciones",
            "description": "Letras de opciones: 'ABCDE' o ['A','B','C'] (2-5 opciones)",
            "default": "ABCDE"
          }
        },
        "type": "object",
        "required": [
          "identificador",
          "cantidad_preguntas"
        ],
        "title": "GenerarGenericaRequest"
      },
      "GenerarGenericaResponse": {
        "properties": {
          "hoja_id": {
            "type": "string",
            "format": "uuid",
            "title": "Hoja Id"
          },
          "identificador": {
            "type": "string",
            "title": "Identificador"
          },
          "num_opciones": {
            "type": "integer",
            "title": "Num Opciones",
            "description": "Cantidad de burbujas por pregunta (2-5)"
          },
          "qr_data": {
            "type": "object",
            "title": "Qr Data",
            "description": "JSON embebido en el QR, p. ej. {\"tipo\": \"generica\", \"identificador\": \"GEN-001\"}"
          },
          "pdf_base64": {
            "type": "string",
            "title": "Pdf Base64",
            "description": "PDF de la hoja de respuestas en base64"
          }
        },
        "type": "object",
        "required": [
          "hoja_id",
          "identificador",
          "num_opciones",
          "qr_data",
          "pdf_base64"
        ],
        "title": "GenerarGenericaResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LecturaGenericaItem": {
        "properties": {
          "identificador": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identificador"
          },
          "identificador_persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identificador Persona"
          },
          "cantidad_preguntas": {
            "type": "integer",
            "title": "Cantidad Preguntas"
          },
          "num_opciones": {
            "type": "integer",
            "title": "Num Opciones"
          },
          "respuestas": {
            "items": {
              "$ref": "#/components/schemas/RespuestaLecturaItem"
            },
            "type": "array",
            "title": "Respuestas"
          },
          "imagen_anotada_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imagen Anotada Base64"
          },
          "qr_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Qr Codes"
          }
        },
        "type": "object",
        "required": [
          "cantidad_preguntas",
          "num_opciones",
          "respuestas",
          "qr_codes"
        ],
        "title": "LecturaGenericaItem"
      },
      "LecturaGenericaLoteResponse": {
        "properties": {
          "total_paginas": {
            "type": "integer",
            "title": "Total Paginas"
          },
          "paginas_leidas": {
            "type": "integer",
            "title": "Paginas Leidas"
          },
          "paginas_ignoradas": {
            "type": "integer",
            "title": "Paginas Ignoradas"
          },
          "lecturas": {
            "items": {
              "$ref": "#/components/schemas/LecturaGenericaItem"
            },
            "type": "array",
            "title": "Lecturas"
          },
          "errores": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Errores",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "total_paginas",
          "paginas_leidas",
          "paginas_ignoradas",
          "lecturas"
        ],
        "title": "LecturaGenericaLoteResponse"
      },
...
      "RespuestaLecturaItem": {
        "properties": {
          "pregunta": {
            "type": "integer",
            "title": "Pregunta"
          },
          "opciones_seleccionadas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Opciones Seleccionadas"
          }
        },
        "type": "object",
        "required": [
          "pregunta",
          "opciones_seleccionadas"
        ],
        "title": "RespuestaLecturaItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}